don't forget RT length when switching between chars. fixes WAR-396
git-svn-id: http://svn.warlock.cc/svn/warlock/warlock2/trunk@2044 bb130017-240f-0410-9613-9c2c827820fb

+++ stormfront/cc.warlock.core.stormfront/src/main/cc/warlock/core/stormfront/internal/PromptTagHandler.java
@@ -53,7 +53,7 @@ public class PromptTagHandler extends DefaultTagHandler {
 		prompt.setLength(0);
 		
 		if (attributes.getValue("time") != null)
-			handler.getClient().syncTime(new Integer(attributes.getValue("time")));
+			handler.getClient().syncTime(new Long(attributes.getValue("time")));
 	}
 	
 	@Override
